fix(web): align conversation find with rendered messages
